Move utility classes and functions to separate header files
